+++ CHANGELOG.md
@@ -2,14 +2,14 @@
 
 ## **5.2.2** (in progress)
 
+### Features added:
+* Make UI changes to the Data Import landing page (UIDATIMP-1218)
+
 ### Bugs fixed:
 * View all logs: filters are not updated after logs deletion (UIDATIMP-1219)
 * View all logs: User and Job filter contains only 10 records (UIDATIMP-1220)
 
 ## [5.2.1](https://github.com/folio-org/ui-data-import/tree/v5.2.1) (2022-07-27)
-
-### Features added:
-* Make UI changes to the Data Import landing page (UIDATIMP-1218)
 
 ### Bugs fixed:
 * Long titles do not fit in the confirmation modal window header (UIDATIMP-1196)